requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
fastapi==0.115.4
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
greenlet==3.1.1
groq==0.11.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.4
langchain-core==0.3.14
langchain-groq==0.2.0
langchain-text-splitters==0.3.0
langgraph==0.2.41
langgraph-checkpoint==2.0.2
langgraph-sdk==0.1.35
langserve==0.3.0
langsmith==0.1.137
msgpack==1.1.0
multidict==6.1.0
numpy==1.26.4
orjson==3.10.10
packaging==24.1
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.4.5
setuptools==75.1.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sse-starlette==1.8.2
starlette==0.41.2
tenacity==9.0.0
tokenizers==0.20.1
tqdm==4.66.6
transformers==4.46.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
wheel==0.44.0
yarl==1.16.0
